fix(tap): keep the standalone shim graph react-free in dist and run the alias e2e in ci
