+++ packages/tap/vitest.config.ts
@@ -1,4 +1,8 @@
+import { resolve } from "node:path";
+import { fileURLToPath } from "node:url";
 import { defineConfig } from "vitest/config";
+
+const packageRoot = fileURLToPath(new URL(".", import.meta.url));
 
 export default defineConfig({
   test: {
@@ -8,6 +12,31 @@ export default defineConfig({
           name: "dev",
           environment: "jsdom",
           include: ["src/**/*.test.{ts,tsx}"],
+          globals: true,
+        },
+      },
+      {
+        // Proves the react-free path: `react` resolves to the standalone shim,
+        // so these tests fail if real React sneaks back into the shim's graph.
+        resolve: {
+          alias: [
+            {
+              find: /^react\/compiler-runtime$/,
+              replacement: resolve(
+                packageRoot,
+                "src/standalone-shim/compiler-runtime.ts",
+              ),
+            },
+            {
+              find: /^react$/,
+              replacement: resolve(packageRoot, "src/standalone-shim/index.ts"),
+            },
+          ],
+        },
+        test: {
+          name: "standalone",
+          environment: "node",
+          include: ["src/__tests__/**/*.e2e.ts"],
           globals: true,
         },
       },
